Updated to v0.1
Finally working!

Took out some code, moved some around. No need for functionality that isn't added yet.

Had some bizarre problems with Tkinter and issues with having the declaration of parts of the window on the same line as the pack() method. So I split some lines, and it fixed it, did the same with sendBtn.